requirements.txt
asgiref==3.11.0
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
